Adds update and remove functionality
The real juice is in src/lib/utils.js

This commit introductes two new functions, partial and pipe, which
together introduce new patterns and allow for a cleaner syntax and
flow

Partial binds an argument to a function. This means that functions
can be pre-populated with arguments, like id, that are needed but
less relevant to the situation at hand.

Pipe allows easy, readable function chaining, which is necessary
and makes 'one-use' variables a thing of the past.

This commit uses these new functions to add handleToggle and
handleRemove to update state in the App, TodoList, and TodoItem
components