Adds simple Router to app
This commit adds three components: Footer, Link, Router that
will soon allow the todos to be filtered by categories.

The link and router work hand in hand to update the app's context
with the proper url and history. This allows a browser's navigation
to populate and function without a page reload