Adds persistence to app through json-server
Until this commit any data was only saved in the browser's memory. This
is a problem if, in theory, the todo list was important.

Using node's json-server which can be installed globally with
`npm i -g json-server` GET, POST, PUT and DELETE methods are sprinkled
into the App as they ought to be

To see it in action fire up the server locally with:

`json-server -p 8080 --watch db.json`

Flash temp messages are displayed in the interface. No error handling
was incorporated into the application

Changes are stored in the project's db.json file, found at the root